MAR-3563. Added internal info to get quora metric throug config
