+++ src/quora/interfaces/FirestoreData.ts
@@ -1,0 +1,17 @@
+export interface FirestoreData {
+  data: [
+    {
+      document: {
+        name: string
+        fields: {
+          count: {
+            integerValue: string
+          }
+        },
+        createTime: string
+        updateTime: string
+      },
+      readTime: string
+    }
+  ]
+}
