+++ src/quora/interfaces/QuoraParams.ts
@@ -1,0 +1,3 @@
+export interface QuoraParams {
+  quoraUserID: string;
+}
